requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
ecdsa==0.19.0
email-validator==2.1.0
fastapi==0.109.2
h11==0.14.0
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.36.3
typing_extensions==4.12.2
uvicorn==0.27.1
wheel==0.45.0
stripe
websockets